Create separate file
Create separate class and enum file
